+++ src/doc.ts
@@ -17,6 +17,7 @@ const markdowns = {
     main: readFileSync(join(__dirname, "../readme.md"), {encoding: "utf-8"}),
     help: readFileSync(join(__dirname, "../assets/docs/help.md"), {encoding: "utf-8"}),
     syntax: readFileSync(join(__dirname, "../assets/docs/syntax.md"), {encoding: "utf-8"}),
+    changelog: readFileSync(join(__dirname, "../changelog.md"), {encoding: "utf-8"}),
 }
 
 export function render(md: string, variables: {[key: string]: string} = {}, var_prefix = "<:", var_suffix = ">"): void {
@@ -25,13 +26,14 @@ export function render(md: string, variables: {[key: string]: string} = {}, var_
     }
 
     const content = marked.parse(md)
-    console.info(content)
+    console.info(content.trim())
 }
 export default function accessToDoc(args: Arguments, props: DotProperties): void {    
     if(args.getOptionValue('help')) return render(markdowns.help)
+    else if(args.getOptionValue('changelog')) return render(markdowns.changelog)
     else if(args.getOptionValue('syntax')) return render(markdowns.syntax)
     else if(args.getOptionValue('info')) return render(markdowns.main)
     else if(args.getOptionValue('props')) return render(props.getProperties.map(({key, value}, index) => `[${index + 1}]> *${key}* = **${value}**`).join('\r\n'))
-    else if(args.getOptionValue('version')) return console.info(version)
+    else if(args.getOptionValue('version')) return render(`*${version}*`)
     else console.error("ERROR => Invalid argument.");
 }