+++ src/parser/objects/comment.ts
@@ -14,8 +14,8 @@ export type CommentaryReturn = {
 const opennersReg = RegExp_OR(rules.comments.openner)
 const closerReg = RegExp_OR(rules.comments.closer)
 export const CommentaryRegexps = {
-    fast: new RegExp(`(?:${opennersReg.source}{2}.*?$)|(?:${opennersReg.source}(?:.|\\s)+?${closerReg.source})`, "m"),
-    detailed: new RegExp(`(?:${opennersReg.source}{2}\\s*(?<inlineContent>.*?)\\s*$)|(?:${opennersReg.source}\\s*(?<blockContent>(?:.|\\s)+?)\\s*${closerReg.source})`, "m")
+    fast: new RegExp(`(?:${opennersReg.source}{2}.*?$)|(?:${opennersReg.source}[^${rules.comments.closer.join('')}]+?${closerReg.source})`, "m"),
+    detailed: new RegExp(`(?:${opennersReg.source}{2}\\s*(?<inlineContent>.*?)\\s*$)|(?:${opennersReg.source}\\s*(?<blockContent>[^${rules.comments.closer.join('')}]+?)\\s*${closerReg.source})`, "m")
 }
 
 export default class Commentary extends CompilerObject {